web/backend/requirements.txt
fastapi
uvicorn
sqlalchemy>=2.0
slowapi>=0.1.9
httpx>=0.27.0
python-multipart
PyJWT>=2.8.0
bcrypt>=4.1.0
pydantic-settings>=2.0.0
email-validator>=2.0.0

# PDF motoru (proje kökündeki src.pdf_engine): /api/pdf-to-word, sıkıştırma vb.
pikepdf>=10
pymupdf>=1.24
pdf2docx>=0.5
pypdf>=6
python-docx
Pillow
lxml
PyPDF2
pdf2image
openpyxl
reportlab
pdfplumber
